backend/requirements.txt
# FastAPI Backend Dependencies

# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & ORM
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
psycopg2-binary>=2.9.0

# Geospatial Data Processing
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0  # For coordinate transformations

# GURS Data Processing
gdal>=3.7.0  # For SHP file processing
fiona>=1.9.5  # Easier interface to GDAL

# Configuration Management
python-dotenv>=1.0.0

# Fuzzy string matching for settlement names
python-Levenshtein>=0.21.0
fuzzywuzzy>=0.18.0

# HTTP client for external APIs
httpx>=0.25.0

# CORS
python-multipart>=0.0.6

# Utilities
python-json-logger>=2.0.7
